corrected/improved deleteProduct

+++ backend/controllers/product.controller.js
@@ -73,6 +73,13 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({
+      success: false,
+      message: "Invalid Product ID",
+    });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({
@@ -81,9 +88,9 @@ export const deleteProduct = async (req, res) => {
     });
   } catch (error) {
     console.log("Error in deleting product:", error.message);
-    res.status(404).json({
+    res.status(500).json({
       success: true,
-      message: "Product Not Found!",
+      message: "Server Error",
     });
   }
 };
